Added the mongodb connection as Rocket::State
Rocket is set up to handle some dbs already, like SQLite. It is not set up to handle mongodb. However it provides an easy way to set up new db pools that use the rust r2d2 package as a gerneric connection pool manager. All well and good, *except* that for some reason mongodb shouldn't use it? There was a package called r2d2-mongodb that did this but it has been deprecated in some way. I do not understand how you are supposed to do this then.

So the next option is to use mongodb directly through the Rust concept of  which on first glance seems like a dependency injection system. So I hope that works?

I need to organize all this code in a much less messy way