Huge problems here. I have no idea how to solve it.
To summarize: when I try and use the db connection to get the books, I recieve a panic here: https://github.com/mongodb/mongo-rust-driver/blob/v2.1.0/src/cmap/connection_requester.rs#L39 which is fun because the comment around it is that 'this will never happen' so ah. Maybe I'm not meant to keep connections around? I am going to try and switch to a mongopool